refactor(java): share runtime core for concore and concoredocker
